# Optimize `System.Reflection.Metadata.ByteSequenceComparer`.