--- a/src/libraries/System.Reflection.Metadata/tests/Utilities/HashTests.cs
+++ b/src/libraries/System.Reflection.Metadata/tests/Utilities/HashTests.cs
@@ -18,7 +18,7 @@ namespace System.Reflection.Metadata.Tests
         [Fact]
         public void GetFNVHashCodeImmutableByteTest()
         {
-            Assert.Equal(-1088511923, Hash.GetFNVHashCode(ImmutableArray.Create((byte)0xFF, (byte)0xD1)));
+            Assert.Equal(-1088511923, Hash.GetFNVHashCode(ImmutableArray.Create((byte)0xFF, (byte)0xD1).AsSpan()));
         }
 
         [Fact]