This PR uses the optimized `MemoryExtensions.SequenceEqual` in place of simple `for` loops over the arrays~~, and also uses `System.HashCode.AddBytes` instead of the custom FNV hash implementation on frameworks that support it~~.

I benchmarked the hash code changes, it's slower on tiny arrays but it's getting better at lengths after around 100 bytes.

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1706 (21H2)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.100-preview.5.22267.11
  [Host]     : .NET 7.0.0 (7.0.22.26611), X64 RyuJIT
  DefaultJob : .NET 7.0.0 (7.0.22.26611), X64 RyuJIT


```
|   Method |    Size |              Mean |          Error |         StdDev |
|--------- |-------- |------------------:|---------------:|---------------:|
| **HashCode** |       **1** |         **7.2412 ns** |      **0.1760 ns** |      **0.1808 ns** |
|      Fnv |       1 |         0.1868 ns |      0.0419 ns |      0.0392 ns |
| **HashCode** |      **10** |        **15.8134 ns** |      **0.2868 ns** |      **0.2682 ns** |
|      Fnv |      10 |         4.6705 ns |      0.1316 ns |      0.1711 ns |
| **HashCode** |     **100** |        **75.6003 ns** |      **1.3908 ns** |      **1.2329 ns** |
|      Fnv |     100 |       102.0950 ns |      1.1608 ns |      1.0290 ns |
| **HashCode** |    **1000** |       **750.8242 ns** |     **12.9713 ns** |     **11.4987 ns** |
|      Fnv |    1000 |     1,219.6150 ns |      6.8126 ns |      5.6888 ns |
| **HashCode** |   **10000** |     **7,163.1200 ns** |    **138.3457 ns** |    **148.0284 ns** |
|      Fnv |   10000 |    12,338.7566 ns |    112.8478 ns |    100.0367 ns |
| **HashCode** |  **100000** |    **71,553.8933 ns** |  **1,400.4046 ns** |  **1,375.3846 ns** |
|      Fnv |  100000 |   124,787.2225 ns |  1,473.3178 ns |  1,378.1423 ns |
| **HashCode** | **1000000** |   **717,269.7070 ns** | **13,429.2555 ns** | **15,465.1506 ns** |
|      Fnv | 1000000 | 1,253,888.7630 ns | 10,577.0581 ns |  9,893.7867 ns |

Here's the benchmark source file:

``` csharp
using System.Security.Cryptography;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkRunner.Run<HashCodeBenchmark>();

public class HashCodeBenchmark
{
    private byte[] _bytes = null!;

    [Params(1, 10, 100, 1000, 10000, 100000, 1000000)]
    public int Size { get; set; }

    [GlobalSetup]
    public void Initialize()
    {
        _bytes = RandomNumberGenerator.GetBytes(Size);
    }

    [Benchmark]
    public int HashCode()
    {
        HashCode hc = default;
        hc.AddBytes(_bytes);
        return hc.ToHashCode();
    }

    [Benchmark]
    public int Fnv()
    {
        byte[] bytes = _bytes;
        int num = -2128831035;
        for (int i = 0; i < bytes.Length; i++)
        {
            num = (num ^ bytes[i]) * 16777619;
        }
        return num;
    }
}
```