@teo-tsirpanis 2022-06-01
> What are the typical lengths used in this library with these functions?

The content of byte arrays is hashed mostly as part of detecting duplicate entries in the Blob Heap. This is the Blob Heap of `System.Private.Corelib`:

![image](https://user-images.githubusercontent.com/12659251/171425194-77031c55-416a-474f-b7b8-30ba0a4fce08.png)

Above the marked line are the blobs with more than 100 bytes, and below are these with less. Judging from the scrollbar's position, the overwhelming majority of the blobs are smaller than 100 bytes. In another assembly, `System.Text.RegularExpressions`, only 11 out of 1899 blobs (0.6%) are smaller.

The matter is a bit more nuanced, since `HashCode`'s xxHash is of better quality than FNV and also randomized (doesn't matter here, the blob heap's contents are not untrusted), which might lead to better `Dictionary` lookup performance.

I ran a quick and likely not correct benchmark to measure the different hash codes interacting with a dictionary, and the results were not promising:

``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1706 (21H2)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.100-preview.5.22267.11
  [Host]     : .NET 7.0.0 (7.0.22.26611), X64 RyuJIT
  DefaultJob : .NET 7.0.0 (7.0.22.26611), X64 RyuJIT
```

|   Method |    Size |             Mean |           Error |          StdDev |
|--------- |-------- |-----------------:|----------------:|----------------:|
| **HashCode** |       **1** |         **278.1 ns** |         **5.03 ns** |         **4.46 ns** |
|      Fnv |       1 |         263.3 ns |         5.11 ns |         4.78 ns |
| **HashCode** |      **10** |       **1,538.2 ns** |        **22.18 ns** |        **20.75 ns** |
|      Fnv |      10 |       1,257.7 ns |        24.44 ns |        25.10 ns |
| **HashCode** |     **100** |      **13,028.9 ns** |       **248.42 ns** |       **408.16 ns** |
|      Fnv |     100 |      10,465.3 ns |       113.62 ns |       106.28 ns |
| **HashCode** |    **1000** |     **130,499.0 ns** |     **1,847.54 ns** |     **1,728.19 ns** |
|      Fnv |    1000 |     104,427.7 ns |     2,006.57 ns |     1,970.72 ns |
| **HashCode** |   **10000** |   **2,503,221.2 ns** |    **48,320.29 ns** |    **45,198.83 ns** |
|      Fnv |   10000 |   2,241,465.7 ns |    39,239.79 ns |    55,008.71 ns |
| **HashCode** |  **100000** |  **25,288,618.9 ns** |   **502,327.05 ns** |   **955,729.43 ns** |
|      Fnv |  100000 |  22,715,999.2 ns |   451,626.01 ns | 1,173,836.04 ns |
| **HashCode** | **1000000** | **361,844,631.2 ns** | **6,858,037.74 ns** | **6,735,510.06 ns** |
|      Fnv | 1000000 | 324,495,645.0 ns | 6,160,111.58 ns | 7,093,993.66 ns |

``` csharp
using System.Diagnostics.CodeAnalysis;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkRunner.Run<HashCodeBenchmark>();

public class HashCodeBenchmark
{
    [Params(1, 10, 100, 1000, 10000, 100000, 1000000)]
    public int Size { get; set; }

    private static byte[] NextBytes(Random rng)
    {
        byte[] bytes = new byte[rng.Next(1, 16)];
        rng.NextBytes(bytes);
        return bytes;
    }

    public int Run(IEqualityComparer<byte[]> comparer)
    {
        Random rng = new();
        byte[][] bytes = new byte[Size][];
        Dictionary<byte[], int> dict = new(comparer);

        for (int i = 0; i < Size; i++)
        {
            byte[] randomBytes = NextBytes(rng);
            bytes[i] = randomBytes;
            if (rng.Next() % 2 == 0)
                dict[randomBytes] = i;
        }

        int acc = 0;
        foreach (var x in bytes)
        {
            if (dict.TryGetValue(x, out int value))
                acc += value;
        }

        return acc;
    }

    [Benchmark]
    public int HashCode() => Run(new HashCodeComparer());

    [Benchmark]
    public int Fnv() => Run(new FnvComparer());
}

public sealed class FnvComparer : IEqualityComparer<byte[]>
{
    public bool Equals(byte[]? x, byte[]? y) => x.AsSpan().SequenceEqual(y);

    public int GetHashCode([DisallowNull] byte[] obj)
    {
        ReadOnlySpan<byte> bytes = obj.AsSpan();
        int num = -2128831035;
        for (int i = 0; i < bytes.Length; i++)
        {
            num = (num ^ bytes[i]) * 16777619;
        }
        return num;
    }
}

public sealed class HashCodeComparer : IEqualityComparer<byte[]>
{
    public bool Equals(byte[]? x, byte[]? y) => x.AsSpan().SequenceEqual(y);

    public int GetHashCode([DisallowNull] byte[] obj)
    {
        HashCode hc = new();
        hc.AddBytes(obj);
        return hc.ToHashCode();
    }
}
```

The best performance measurement would be to run Roslyn before and after this change but that's too much hassle. So unless there's any objection, I will revert the hash code changes.