@teo-tsirpanis 2022-06-20
What to do with the hash code algorithm? Revert it?