@stephentoub 2022-06-29
Yeah, I suggest just culling this back to simplifying Equals.