@teo-tsirpanis 2022-06-29
All right, I rolled back to the old algorithm. And I merged the two overloads that take `byte[]` and `ImmutableArray<byte>` into one that takes a `ReadOnlySpan<byte>`.